score: avoid warnings about type defaulting.

+++ src/Score.hs
@@ -8,7 +8,7 @@ import Graph
 
 -- | squared distance
 sqrDistance :: (GridPoint, GridPoint) -> Int
-sqrDistance ((px, py), (qx, qy)) = (px - qx)^2 + (py - qy)^2
+sqrDistance ((px, py), (qx, qy)) = (px - qx)^(2::Int) + (py - qy)^(2::Int)
 
 dislike :: ([GridPoint], [GridPoint]) -> Int
 dislike (hole, pose) = sum [f h | h <- hole]
@@ -41,6 +41,6 @@ tolerant eps (vi, vj) (vi', vj')
   | d' < d = (c - eps) * d <= c * d'
   | otherwise = True
   where
-    c = 10^6
+    c = 10^(6::Int)
     d  = sqrDistance (vi,  vj )
     d' = sqrDistance (vi', vj')
